// src/enums/agreement.type.ts
export enum AgreementType {
  APPLY_INVESTOR_FL = 'Заявление инвестора (ФЛ)',
  APPLY_INVESTOR_UL = 'Заявление инвестора (ЮЛ, ИП)',
  APPLY_LENDER = 'Заявление ЛПИ',
  LENDER_OFERTA = 'Оферта заемщика',
  INVESTOR_CLAIM = 'Акцепт инвестора',
  INVOICE = 'Счет на оплату',
}